requirements.txt
click==7.1.1
Flask==1.1.2
flask-cors==3.0.7
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyMySQL==0.9.3
pymysql-pooling==1.0.6
waitress==1.4.3
Werkzeug==1.0.1
google-auth==1.15.0
requests==2.23.0
